Enable redirect-based OAuth flow
Production usages untested.

+++ src/googleApis/authenticate.js
@@ -1,9 +1,8 @@
-import { GOOGLE_APIS_CLIENT_ID, GOOGLE_APIS_CLIENT_SECRET } from '../constants';
-import path from 'path';
+import { GOOGLE_APIS_CLIENT_ID, GOOGLE_APIS_CLIENT_SECRET } from '../constants'; // don't forget to set the secret!
 
-const readline = require('readline');
-const { google } = require('googleapis');
-const { backgroundAccessTokens } = require('./../database');
+import { google } from 'googleapis';
+import { createServer } from 'http'; // built-in node HTTP webserver
+import { backgroundAccessTokens } from './../database'; // corresponds to DB table
 
 const SCOPES = [
 	'https://www.googleapis.com/auth/calendar https://mail.google.com/ email profile'
@@ -12,7 +11,7 @@ const SCOPES = [
 const oAuth2Client = new google.auth.OAuth2(
 	GOOGLE_APIS_CLIENT_ID,
 	GOOGLE_APIS_CLIENT_SECRET,
-	'urn:ietf:wg:oauth:2.0:oob'
+	'http://localhost:3001'
 );
 
 const authUrl = oAuth2Client.generateAuthUrl({
@@ -22,26 +21,32 @@ const authUrl = oAuth2Client.generateAuthUrl({
 
 console.log('Authorize this app by visiting this url:', authUrl);
 
-const rl = readline.createInterface({
-	input: process.stdin,
-	output: process.stdout
-});
+// launch local HTTP server for Google OAuth redirect
+createServer(function (req, res) {
+	const params = new URLSearchParams(req.url.slice(2)); // URL parser
+	if (params.has('code')) {
+		// we got a callback from the OAuth flow
+		oAuth2Client.getToken(params.get('code')).then(async ({ tokens }) => {
+			// out with the old...
+			await backgroundAccessTokens.destroy({
+				where: {
+					service: 'google'
+				}
+			});
+			// in with the new!
+			await backgroundAccessTokens.create({
+				service: 'google',
+				token: JSON.stringify(tokens)
+			});
 
-rl.question('Enter the code from that page here: ', code => {
-	rl.close();
-	oAuth2Client.getToken(code).then(async res => {
-		const token = res.tokens;
-		await backgroundAccessTokens.destroy({
-			where: {
-				service: 'google'
-			}
+			console.log('Authentication complete! - Now you can run the API!');
+			res.write('Authentication Complete!');
+			res.end();
+			process.exit(0);
 		});
-
-		await backgroundAccessTokens.create({
-			service: 'google',
-			token: JSON.stringify(token)
-		});
-
-		console.log('Authentication complete!');
-	});
-});
+	} else {
+		res.writeHead(200);
+		res.write('StuyActivities API Authentication - Calendar and Email');
+		res.end();
+	}
+}).listen(3001);
